// apps/backend/src/inventory/enums/referencia-tipo.enum.ts
export enum ReferenciaTipo {
  factura = 'factura',
  pedido = 'pedido',
  importacion = 'importacion',
  manual = 'manual',
  ajuste_inventario = 'ajuste_inventario',
  transferencia_sucursal = 'transferencia_sucursal',
  factura_recibida = 'factura_recibida',
  factura_importada = 'factura_importada',
}